Refactor commands, config and model modules
* Move S3Bucket and related classes to models/s3_bucket module
* Move ImageBuilderStack and related exceptions to models/imagebuilder_resources module

Tests
* Rename cluster/imagebuilder_dummy_model to cluster/dummy_imagebuilder_config
* Move s3 bucket mocking to dummy_s3_bucket
