+++ cli/src/pcluster/cli_commands/configure/easyconfig.py
@@ -25,7 +25,13 @@ from pcluster.cli_commands.configure.networking import (
     automate_subnet_creation,
     automate_vpc_with_subnet_creation,
 )
-from pcluster.cli_commands.configure.utils import get_regions, get_resource_tag, handle_client_exception, prompt, prompt_iterable
+from pcluster.cli_commands.configure.utils import (
+    get_regions,
+    get_resource_tag,
+    handle_client_exception,
+    prompt,
+    prompt_iterable,
+)
 from pcluster.constants import DEFAULT_MAX_COUNT, DEFAULT_MIN_COUNT, SUPPORTED_SCHEDULERS
 from pcluster.utils import default_config_file_path, error, get_region, get_supported_os_for_scheduler
