+++ cli/src/pcluster/models/cluster.py
@@ -25,11 +25,11 @@ from marshmallow import ValidationError
 from common.aws.aws_api import AWSApi
 from common.boto3.common import AWSClientError, StackNotFoundError
 from pcluster.cli_commands.compute_fleet_status_manager import ComputeFleetStatus, ComputeFleetStatusManager
+from pcluster.config.cluster_config import BaseClusterConfig, SlurmScheduling, Tag
 from pcluster.config.config_patch import ConfigPatch
 from pcluster.constants import PCLUSTER_STACK_PREFIX
-from pcluster.config.cluster_config import BaseClusterConfig, SlurmScheduling, Tag
 from pcluster.models.cluster_resources import ClusterInstance, ClusterStack
-from pcluster.config.common import S3Bucket, S3BucketFactory, S3FileFormat
+from pcluster.models.s3_bucket import S3Bucket, S3BucketFactory, S3FileFormat
 from pcluster.schemas.cluster_schema import ClusterSchema
 from pcluster.templates.cdk_builder import CDKTemplateBuilder
 from pcluster.utils import generate_random_name_with_prefix, get_installed_version, get_region, grouper
