+++ cli/src/pcluster/templates/cluster_stack.py
@@ -38,7 +38,6 @@ from aws_cdk.core import (
 )
 
 from common.aws.aws_api import AWSApi
-from pcluster.constants import OS_MAPPING
 from pcluster.config.cluster_config import (
     BaseQueue,
     HeadNode,
@@ -48,7 +47,8 @@ from pcluster.config.cluster_config import (
     SharedStorageType,
     SlurmClusterConfig,
 )
-from pcluster.config.common import S3Bucket
+from pcluster.constants import OS_MAPPING
+from pcluster.models.s3_bucket import S3Bucket
 from pcluster.templates.awsbatch_builder import AwsBatchConstruct
 from pcluster.templates.cdk_builder_utils import (
     PclusterLambdaConstruct,
