+++ cli/src/pcluster/templates/slurm_builder.py
@@ -17,9 +17,9 @@ from aws_cdk import aws_logs as logs
 from aws_cdk import aws_route53 as route53
 from aws_cdk.core import CfnOutput, CfnTag, Construct, CustomResource, Fn, Stack
 
-from pcluster.constants import OS_MAPPING
 from pcluster.config.cluster_config import CapacityType, SharedStorageType, SlurmClusterConfig
-from pcluster.config.common import S3Bucket
+from pcluster.constants import OS_MAPPING
+from pcluster.models.s3_bucket import S3Bucket
 from pcluster.templates.cdk_builder_utils import (
     PclusterLambdaConstruct,
     add_lambda_cfn_role,
