+++ cli/tests/pcluster/models/test_cluster.py
@@ -9,7 +9,6 @@
 # OR CONDITIONS OF ANY KIND, express or implied. See the License for the specific language governing permissions and
 # limitations under the License.
 import json
-from typing import List
 from unittest.mock import PropertyMock
 
 import pytest
@@ -18,153 +17,15 @@ from assertpy import assert_that
 
 from common.boto3.common import AWSClientError
 from pcluster.models.cluster import Cluster, ClusterActionError, NodeType
-from pcluster.config.cluster_config import Resource, Tag
+from pcluster.config.cluster_config import Tag
 from pcluster.models.cluster_resources import ClusterStack
-from pcluster.config.common import S3Bucket
-from pcluster.validators.common import FailureLevel, Validator
+from pcluster.models.s3_bucket import S3Bucket
 from tests.common.dummy_aws_api import mock_aws_api
-from tests.pcluster.models.cluster_dummy_model import (
-    dummy_slurm_cluster_config,
-    mock_bucket,
-    mock_bucket_object_utils,
-    mock_bucket_utils,
-)
+from tests.pcluster.config.dummy_cluster_config import dummy_slurm_cluster_config
+from tests.pcluster.models.dummy_s3_bucket import mock_bucket, mock_bucket_object_utils, mock_bucket_utils
 from tests.pcluster.test_utils import FAKE_CLUSTER_NAME, FAKE_STACK_NAME
 
 LOG_GROUP_TYPE = "AWS::Logs::LogGroup"
-
-
-class FakeInfoValidator(Validator):
-    """Dummy validator of info level."""
-
-    def _validate(self, param):
-        self._add_failure(f"Wrong value {param}.", FailureLevel.INFO)
-
-
-class FakeErrorValidator(Validator):
-    """Dummy validator of error level."""
-
-    def _validate(self, param):
-        self._add_failure(f"Error {param}.", FailureLevel.ERROR)
-
-
-class FakeComplexValidator(Validator):
-    """Dummy validator requiring multiple parameters as input."""
-
-    def _validate(self, fake_attribute, other_attribute):
-        self._add_failure(f"Combination {fake_attribute} - {other_attribute}.", FailureLevel.WARNING)
-
-
-class FakePropertyValidator(Validator):
-    """Dummy property validator of info level."""
-
-    def _validate(self, property_value: str):
-        self._add_failure(f"Wrong value {property_value}.", FailureLevel.INFO)
-
-
-def _assert_validation_result(result, expected_level, expected_message):
-    """Assert that validation results is the expected one, by checking level and message."""
-    assert_that(result.level).is_equal_to(expected_level)
-    assert_that(result.message).contains(expected_message)
-
-
-def test_resource_validate():
-    """Verify that validators are executed in the right order according to priorities with the expected params."""
-
-    class FakeResource(Resource):
-        """Fake resource class to test validators."""
-
-        def __init__(self):
-            super().__init__()
-            self.fake_attribute = "fake-value"
-            self.other_attribute = "other-value"
-
-        def _validate(self):
-            self._execute_validator(FakeErrorValidator, param=self.fake_attribute)
-            self._execute_validator(
-                FakeComplexValidator,
-                fake_attribute=self.fake_attribute,
-                other_attribute=self.other_attribute,
-            )
-            self._execute_validator(FakeInfoValidator, param=self.other_attribute)
-
-    fake_resource = FakeResource()
-    validation_failures = fake_resource.validate()
-
-    # Verify high prio is the first of the list
-    _assert_validation_result(validation_failures[0], FailureLevel.ERROR, "Error fake-value.")
-    _assert_validation_result(validation_failures[1], FailureLevel.WARNING, "Combination fake-value - other-value.")
-    _assert_validation_result(validation_failures[2], FailureLevel.INFO, "Wrong value other-value.")
-
-
-def test_dynamic_property_validate():
-    """Verify that validators of dynamic parameters are working as expected."""
-
-    class FakeResource(Resource):
-        """Fake resource class to test validators."""
-
-        def __init__(self):
-            super().__init__()
-            self.deps_value = ""
-
-        def _validate(self):
-            self._execute_validator(FakePropertyValidator, property_value=self.dynamic_attribute)
-
-        @property
-        def dynamic_attribute(self):
-            return f"dynamic-value: {self.deps_value}"
-
-    fake_resource = FakeResource()
-    validation_failures = fake_resource.validate()
-    _assert_validation_result(
-        validation_failures[0], FailureLevel.INFO, f"Wrong value dynamic-value: {fake_resource.deps_value}."
-    )
-
-    fake_resource.deps_value = "test1"
-    validation_failures = fake_resource.validate()
-    _assert_validation_result(
-        validation_failures[0], FailureLevel.INFO, f"Wrong value dynamic-value: {fake_resource.deps_value}."
-    )
-
-    fake_resource.deps_value = "test2"
-    validation_failures = fake_resource.validate()
-    _assert_validation_result(
-        validation_failures[0], FailureLevel.INFO, f"Wrong value dynamic-value: {fake_resource.deps_value}."
-    )
-
-
-def test_nested_resource_validate():
-    """Verify that validators of nested resources are executed correctly."""
-
-    class FakeNestedResource(Resource):
-        """Fake nested resource class to test validators."""
-
-        def __init__(self, fake_value):
-            super().__init__()
-            self.fake_attribute = fake_value
-
-        def _validate(self):
-            self._execute_validator(FakeErrorValidator, param=self.fake_attribute)
-
-    class FakeParentResource(Resource):
-        """Fake resource class to test validators."""
-
-        def __init__(self, nested_resource: FakeNestedResource, list_of_resources: List[FakeNestedResource]):
-            super().__init__()
-            self.fake_resource = nested_resource
-            self.other_attribute = "other-value"
-            self.list_of_resources = list_of_resources
-
-        def _validate(self):
-            self._execute_validator(FakeInfoValidator, param=self.other_attribute)
-
-    fake_resource = FakeParentResource(FakeNestedResource("value1"), [FakeNestedResource("value2")])
-    validation_failures = fake_resource.validate()
-
-    # Verify children failures are executed first
-    _assert_validation_result(validation_failures[0], FailureLevel.ERROR, "Error value1.")
-    _assert_validation_result(validation_failures[1], FailureLevel.ERROR, "Error value2.")
-    _assert_validation_result(validation_failures[2], FailureLevel.INFO, "Wrong value other-value.")
 
 
 class TestCluster:
