+++ cli/tests/pcluster/templates/test_cw_dashboard_builder.py
@@ -19,8 +19,7 @@ from pcluster.config.cluster_config import SharedStorageType
 from pcluster.schemas.cluster_schema import ClusterSchema
 from pcluster.templates.cdk_builder import CDKTemplateBuilder
 from tests.common.dummy_aws_api import mock_aws_api
-
-from ..models.cluster_dummy_model import dummy_cluster_bucket, mock_bucket
+from tests.pcluster.models.dummy_s3_bucket import dummy_cluster_bucket, mock_bucket
 
 
 @pytest.mark.parametrize(
